+++ lib/core/utils/constants.dart
@@ -1,4 +1,4 @@
-import 'package:videos/presentation/not_found_page.dart';
+import 'package:videos/screens/not_found_page.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
