+++ lib/screens/home_details_screen.dart
